splits fruit, html_helper and layout
