+++ app/admin/arbre_helpers.rb
@@ -1,82 +1,4 @@
 module ArbreHelpers
-  def self.fruit_show_page(context)
-    context.instance_eval do
-      columns do
-        column span: 2 do
-          ArbreHelpers.fruit_attribute_table(self, resource)
-          if resource.respond_to?(:external_link) && !resource.external_link.blank?
-            h4 "External links"
-            ArbreHelpers.show_links(self, resource.external_link.split(',').compact)
-          end 
-          if resource.respond_to?(:extra_info)  && !resource.extra_info.nil?
-            h4 "Extra info"
-            begin 
-              if resource.extra_info
-                extra_info_as_json = JSON.parse(resource.extra_info)
-                ArbreHelpers.extra_info_renderer(self, extra_info_as_json)
-              end
-            rescue JSON::ParserError
-              span resource.extra_info
-            end
-          end
-          if attachments = resource.attachments.presence
-            h3 "Attachments"
-            ArbreHelpers::Attachment.attachments_list(self, attachments)
-          end
-        end
-
-        column do 
-          ArbreHelpers.fruit_relations_panels(self, resource)
-        end 
-      end  
-    end
-  end
-
-  def self.fruit_attribute_table(context, resource, &block)
-    context.instance_eval do 
-      if block
-        attributes_table_for(resource, &block)
-      else
-        blacklist = %i(id person issue_id created_at updated_at replaced_by extra_info external_link)
-        attributes_table_for(resource, *(default_attribute_table_rows - blacklist))
-      end
-    end
-  end
-
-  def self.fruit_relations_panels(context, resource)
-    context.instance_eval do
-      person = resource.person
-
-      attributes_table_for(resource) do
-        row :id
-        if resource.replaced_by
-          row :replaced_by
-        end
-        row :person
-        row :issue
-        row :created_at
-      end
-
-      if previous = resource.previous_versions.presence
-        h3 "Previous versions"
-        previous.each do |r|
-          span r.created_at.strftime("%e %b %Y :")
-          span link_to "#{r.name}", r
-          br
-        end
-      end
-
-      if others = resource.others_for_person.presence
-        h3 "Other #{resource.class.name.pluralize.titleize} for person"
-        others.each do |r|
-          span r.created_at.strftime("%e %b %Y :")
-          span link_to "#{r.name}", r
-          br
-        end
-      end
-    end
-  end
-
   def self.fields_for_replaces(context, form, assoc)
     Appsignal.instrument("render_fields_for_replaces") do
       context.instance_eval do
@@ -108,99 +30,11 @@ module ArbreHelpers
     end
   end
 
-  def self.show_links(context, links)
-    context.instance_eval do 
-      div class: 'external_links' do
-        links.each_with_index do |link, index|
-          span link_to("Link ##{index + 1}",
-            link,
-            target: '_blank'
-          )
-          br
-        end
-      end
-    end
-  end
-
-  def self.has_many_links(context, builder, links, title)
-    context.instance_eval do
-      unless links.blank?
-        builder.template.concat("<li class='external_links'>".html_safe) 
-        builder.template.concat("<label>#{title}</label><br />".html_safe)
-        links.each_with_index do |link, index|
-          builder.template.concat( 
-            context.link_to("Link ##{index + 1}",
-              link,
-              target: '_blank'
-            )
-          )
-          builder.template.concat('<br/>'.html_safe)
-        end
-        builder.template.concat('</li>'.html_safe)
-      end 
-    end
-  end
-
-  def self.fruit_collection_show_tab(context, title, relation)
-    Appsignal.instrument("render_#{relation.to_s}") do
-      context.instance_eval do
-        all = resource.send(relation).current.order("created_at DESC")
-        tab "#{title} (#{all.count})" do
-          ArbreHelpers.panel_grid(self, all) do |d|
-            ArbreHelpers.fruit_show_section(self, d)
-          end
-        end
-      end
-    end
-  end
-
-  def self.fruit_show_section(context, fruit, others = [])
-    Appsignal.instrument("render_#{fruit.class.name}") do
-      context.instance_eval do
-        columns = fruit.class.columns.map(&:name) - others.map(&:to_s)
-        columns = columns.map{|c| c.gsub(/_id$/,'') } -
-          %w(id person issue created_at updated_at replaces extra_info external_link)
-        attributes_table_for fruit do
-          row(:show){|o| link_to o.name, o }
-          columns.each do |n|
-            row(n)
-          end
-          others.each do |o|
-            row(o)
-          end
-          if fruit.replaces
-            row(:replaces)
-          end
-          row(:created_at)
-          row(:issue)
-        end
-        if fruit.respond_to?(:external_link) && !fruit.external_link.blank?
-          h4 "External links"
-          ArbreHelpers.show_links(self, fruit.external_link.split(',').compact)
-        end
-        if fruit.respond_to?(:extra_info)  && !fruit.extra_info.nil?
-          h4 "Extra info"
-          begin 
-            if fruit.extra_info
-              extra_info_as_json = JSON.parse(fruit.extra_info)
-              ArbreHelpers.extra_info_renderer(self, extra_info_as_json)
-            end
-          rescue JSON::ParserError
-            span fruit.extra_info
-          end
-        end
-        fruit.attachments.each do |a|
-          ArbreHelpers::Attachment.preview(self, a)
-        end
-      end
-    end
-  end
-
   def self.seed_collection_show_tab(context, title, relation)
     Appsignal.instrument("render_#{relation.to_s}") do
       context.instance_eval do
         tab "#{title} (#{resource.send(relation).count})" do
-          ArbreHelpers.panel_grid(self, resource.send(relation)) do |d|
+          ArbreHelpers::Layout.panel_grid(self, resource.send(relation)) do |d|
             ArbreHelpers.seed_show_section(self, d)
           end
         end
@@ -243,129 +77,5 @@ module ArbreHelpers
     context.instance_eval do
       attributes_table_for resource, :fruit, *columns, *others
     end
-  end
-
-  def self.extra_info_renderer(context, data)
-    context.instance_eval do |ctx|
-      div class: 'extra_info' do 
-        ArbreHelpers.render_extra_info_list(ctx, data)
-      end
-    end
-  end
-
-  def self.render_extra_info_list(context, data)    
-    if data.is_a?(Array)
-      ArbreHelpers.render_extra_info_array(context, data)
-    else 
-      ArbreHelpers.render_extra_info_hash(context, data)
-    end
-  end
-
-  def self.render_extra_info_array(context, data)
-    context.instance_eval do
-      data.each do |value|
-        if ArbreHelpers.is_a_list?(value)
-          hr
-          ArbreHelpers.render_extra_info_list(context, value)
-        else 
-          div do
-            ArbreHelpers.render_link_or_text(context, nil, value.to_s)
-          end
-        end
-      end
-    end
-  end
-
-  def self.render_extra_info_hash(context, data)
-    data.keys.each do |key|
-      label = key
-      value = data[key]
-      if ArbreHelpers.is_a_list?(value)  
-        ArbreHelpers.render_extra_info_list(context, value)
-      else 
-        ArbreHelpers.render_link_or_text(context, label, value.to_s)
-      end
-    end
-  end
-
-  def self.json_renderer(context, data)
-    context.instance_eval do |ctx|
-      context.concat("<li class='extra_info'>".html_safe) 
-      context.concat('<h4>Extra info</h4>'.html_safe) 
-      ArbreHelpers.render_list(ctx, data)
-      context.concat('</li>'.html_safe)
-    end
-  end
-
-  def self.render_list(context, data)
-    if data.is_a?(Array)
-      ArbreHelpers.render_array(context, data)
-    else 
-      ArbreHelpers.render_hash(context, data)
-    end
-  end
-
-  def self.is_a_list?(data)
-    data.is_a?(Array) || data.is_a?(Hash)
-  end
-
-  def self.render_array(context, data)
-    context.concat('<ul>'.html_safe)
-    data.each do |value|
-      if ArbreHelpers.is_a_list?(value)
-        context.concat('<hr/>'.html_safe)
-        value = ArbreHelpers.render_list(context, value)
-      else
-        context.concat('<li>'.html_safe) 
-        ArbreHelpers.render_text_or_link(context, nil, value.to_s)
-        context.concat('</li>'.html_safe)
-      end
-    end
-    context.concat('</ul>'.html_safe) 
-  end
-
-  def self.render_hash(context, data)
-    context.concat('<li>'.html_safe) 
-    data.keys.each do |key|
-      label = key
-      value = data[key]
-      if ArbreHelpers.is_a_list?(value)  
-        ArbreHelpers.render_list(context, value)
-      else
-        ArbreHelpers.render_text_or_link(context, label, value.to_s)
-      end
-    end
-    context.concat('</li>'.html_safe) 
-  end
-
-  def self.render_text_or_link(context, key, text)
-    context.concat("<strong>#{key}: </strong>".html_safe) if key
-    if text.starts_with?('http') || text.starts_with?('ftp') || text.starts_with?('https')
-      context.concat("<a href='#{text}' target='_blank'>#{ArbreHelpers.strip_and_truncate(text)}</a><br/>".html_safe) 
-    else
-      context.concat("#{ArbreHelpers.strip_and_truncate(text)}<br/>".html_safe) 
-    end
-  end
-
-  def self.render_link_or_text(context, key, text)
-    context.instance_eval do
-      strong "#{key}: "
-      if text.starts_with?('http') || text.starts_with?('ftp') || text.starts_with?('https')
-        span do
-          link_to ArbreHelpers.strip_and_truncate(text), text, target: "_blank"
-        end
-      else
-        span ArbreHelpers.strip_and_truncate(text)
-      end
-      br
-    end
-  end
-
-  def self.strip_and_truncate(length = 40, text)
-    ArbreHelpers.strip_html_tags(text).truncate(length, omission:'...')
-  end
-
-  def self.strip_html_tags(text)
-    Nokogiri::HTML(text).text
   end
 end
